Trying to make UI look cool

+++ Twippies/Twippies/Assets/Scripts/UI/Stats/ValueStat.cs
@@ -99,7 +99,7 @@ public class ValueStat : Stat {
         }
     }
 
-    public void Populate(int value, int minValue, int maxValue, string label, bool readOnly, string statName)
+    public void Populate(float value, int minValue, int maxValue, string label, bool readOnly, string statName)
     {
         _statType = StatType.Value;
         _name = "value";
